Uninstalls Semantic UI/Moves the css file into a separate folder/Imports properly
