+++ package.json
@@ -5,9 +5,7 @@
   "dependencies": {
     "axios": "^0.19.0",
     "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.88.0"
+    "react-dom": "^16.9.0"
   },
   "scripts": {
     "start": "PORT=3001 react-scripts start",
